fix(core): re-narrow before the preserve_nonzero underflow test (#2720)
validated_float32_scalar_with_ratio returns a built-in float UN-narrowed
(stored = value if type(value) is float else narrowed), so
cumulant_discovery's preserve_nonzero guard tested the float64 value and
was dead for the annotated input type: _require_float32("gamma", 1e-50,
preserve_nonzero=True) was ACCEPTED while np.float64(1e-50) was REJECTED —
same value, opposite outcome, decided by Python type.

Failing-test-first: TestPreserveNonzeroNarrowing pins rejection for both
types and type/value parity across the boundary region. Fix re-narrows at
the guard (float(np.float32(stored)) == 0.0), matching the correct pattern
already used by option_value_duration.py:98-99. The stored return value is
unchanged for all accepted inputs; only the reject set grows, to exactly
the values whose float32 narrowing is zero.

Closes #2720.

+++ alberta_framework/core/cumulant_discovery.py
@@ -106,7 +106,11 @@ def _require_float32(
         upper_inclusive=upper_inclusive,
         positive=positive,
     )
-    if preserve_nonzero and numerator != 0 and stored == 0.0:
+    # Re-narrow before testing: validated_float32_scalar_with_ratio returns a
+    # built-in float UN-narrowed, so testing ``stored`` directly makes this
+    # guard dead for the annotated input type (issue #2720).  The re-narrowing
+    # matches option_value_duration's underflow guard.
+    if preserve_nonzero and numerator != 0 and float(np.float32(stored)) == 0.0:
         raise ValueError(f"{name} must remain nonzero once narrowed to float32")
     return stored
